experimental support for XSalsa20 with Poly1305 encryption
This introduces a new authenticated encryption method using
XSalsa20 with Poly1305 MAC instead of the old blowfish in ecb mode.

It uses a mostly random nonce in front of every message.

8 Bytes of the nonce are used as timestamp in milliseconds since epoch
and messages that appear to be too far in the future or too far in the
past are discard if they were successfully decrypted to mitigate replay.

+- 10 seconds

switch conversation to the new mode by using /setmethod anything.
switch to old blowfish in ecb mode use /delmethod.

It's essential to use a good password!

How to generate one using pwgen to get at least 256 Bits of entropy?

    Let's try an approximation by using compression :-)

    pwgen -s 1048576|xz -9ve -|wc -c

    796718 <- result

    echo "45*8*796718/1048576"|bc

    273 <- result (Bits of entropy)

    recommendation:
    use pwgen -ns1 45 to get a good password because we just hash it
    down to crypto_secretbox_KEYBYTES
    using crypto_generichash (blake2b) without salt.

Details:
https://download.libsodium.org/libsodium/content/secret-key_cryptography/authenticated_encryption.html

 Changes to be committed:
	modified:   CMakeLists.txt
	modified:   src/CMakeLists.txt
	modified:   src/FiSH.c
	modified:   src/FiSH.h
	new file:   src/XSalsa20Poly1305.c
	new file:   src/XSalsa20Poly1305.h